Refactor: Move pages and components to new structure, remove unused files
- Moved CoursePlannerPage and AboutPage components in the pages directory.
- Updated imports in main.jsx to reflect new component structure.
- Moved the CourseInfoSidebar and CoursePickerSidebar components to be in the components directory.
- Cleaned up the usePlannerUIStore to remove unnecessary console logs.
